refactor memory block related utility functions
