+++ elastica/modules/memory_block.py
@@ -2,7 +2,14 @@ __doc__ = """
 This function is a module to construct memory blocks for different types of systems, such as
 Cosserat Rods, Rigid Body etc.
 """
-from elastica.typing import RodType, RigidBodyType, SystemType, SystemIdxType
+from typing import cast
+from elastica.typing import (
+    RodType,
+    RigidBodyType,
+    SystemType,
+    SystemIdxType,
+    BlockSystemType,
+)
 
 from elastica.rod.rod_base import RodBase
 from elastica.rigidbody.rigid_body import RigidBodyBase
@@ -22,19 +29,21 @@ def construct_memory_block_structures(systems: list[SystemType]) -> list[SystemT
 
     """
     _memory_blocks: list[BlockSystemType] = []
-    temp_list_for_cosserat_rod_systems: list[SystemType] = []
-    temp_list_for_rigid_body_systems: list[SystemType] = []
+    temp_list_for_cosserat_rod_systems: list[RodType] = []
+    temp_list_for_rigid_body_systems: list[RigidBodyType] = []
     temp_list_for_cosserat_rod_systems_idx: list[SystemIdxType] = []
     temp_list_for_rigid_body_systems_idx: list[SystemIdxType] = []
 
     for system_idx, sys_to_be_added in enumerate(systems):
 
         if isinstance(sys_to_be_added, RodBase):
-            temp_list_for_cosserat_rod_systems.append(sys_to_be_added)
+            rod_system = cast(RodType, sys_to_be_added)
+            temp_list_for_cosserat_rod_systems.append(rod_system)
             temp_list_for_cosserat_rod_systems_idx.append(system_idx)
 
         elif isinstance(sys_to_be_added, RigidBodyBase):
-            temp_list_for_rigid_body_systems.append(sys_to_be_added)
+            rigid_body_system = cast(RigidBodyType, sys_to_be_added)
+            temp_list_for_rigid_body_systems.append(rigid_body_system)
             temp_list_for_rigid_body_systems_idx.append(system_idx)
 
         elif isinstance(sys_to_be_added, SurfaceBase):
@@ -71,4 +80,4 @@ def construct_memory_block_structures(systems: list[SystemType]) -> list[SystemT
             )
         )
 
-    return _memory_blocks
+    return list(_memory_blocks)
